moving client to internal folder
